should work, haven't tested with actually unmarshalling the request, but that should be done as part of the pagedetail handler test anyway.
