+++ internal/api/handlers/pagedetail/request.go
@@ -4,6 +4,7 @@ import (
 	"encoding/json"
 	"net/http"
 
+	"github.com/Pergamene/project-spiderweb-service/internal/models/pagedetail"
 	"github.com/julienschmidt/httprouter"
 	"github.com/pkg/errors"
 )
@@ -12,9 +13,9 @@ import (
 type UpdatePageDetailRequest struct {
 	PageGUID       string
 	PageDetailGUID string
-	Title          string                   `json:"title"`
-	Summary        string                   `json:"summary"`
-	Partitions     []map[string]interface{} `json:"partitions"`
+	Title          string                 `json:"title"`
+	Summary        string                 `json:"summary"`
+	Partitions     []pagedetail.Partition `json:"partitions"`
 }
 
 // NewUpdatePageDetailRequest extracts the UpdatePageDetailRequest
@@ -23,6 +24,10 @@ func NewUpdatePageDetailRequest(r *http.Request, p httprouter.Params) (UpdatePag
 	err := json.NewDecoder(r.Body).Decode(&request)
 	if err != nil {
 		return request, errors.New("invalid request")
+	}
+	err = pagedetail.UnmarshalPartitions(request.Partitions)
+	if err != nil {
+		return request, errors.New("not valid page partitions")
 	}
 	request.PageGUID = p.ByName(PageIDRouteKey)
 	request.PageDetailGUID = p.ByName(PageDetailIDRouteKey)
